requirements.txt
streamlit
sentence-transformers
faiss-cpu
google-generativeai
